bot: Bump inter-message delays in sequenced commands
Sometimes these commands are sending messages too fast. Most notably,
the "kang" command can malfunction and cause the sticker management bot
to abort the process of adding the sticker before it reacts to the emoji
being sent. As rare as this may be, it has happened in the past.

Bump the delay when mass-forwarding messages as well to avoid
ratelimiting.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>