bot: Increase sticker adding delay
The Stickers bot takes a relatively long time to add stickers after an
image is uploaded. If we send the /done command too soon, the bot breakd
and does not recognize it, which results in subsequent attempts to add
stickers potentially breaking and/or adding them to the wrong pack. Give
the bot more time to process the image to avoid inconsistent states.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>