modules: Add auto-moderation module
Recently, there has been an influx of spambots which join groups and
send a message as fast as possible before other bots can mute the user.
This is a module that attempts to mitigate this issue by automatically
banning such users and deleting their message if the message has been
sent within 5 seconds (configurable) after joining the group and the
contents of the message are suspicious.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>