text: Add functional character equivalent to zero-width space
telegram-desktop now shows U+200B (zero-width space) as a space, and
Telegram considers it a whitespace character instead of a valid
standalone character. This greatly reduces the useful applications of
the character on Telegram.

The invisible tags in Unicode don't have this problem - Telegram accepts
messages with only a tag character in them and none of the clients
display any visible glyphs when one is present - and have defined
meanings in the Unicode standard. That makes tags perfect for zero-width
space applications.

Use the space tag (U+E0020) as a replacement for the zero-width space
(U+200B) to expand upon the practical applications of empty characters
on Telegram.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>