antibot: Drop most profile checks
Innocent until proven guilty. These checks are prone to false positives
and are unnecessary now that we have sophisticated message content
checks.

The only check left is for "t.me" in the first orlast names, since that
is a blatant form of advertising that does not require sending a
message.

Suggested-by: Artem Labazov <123321artyom@gmail.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>