antibot: Allow forwarding oneself's own messages
Many people will mass-join various support groups in a desperate attempt
to obtain help. In doing so, they forward their single question message
which contains an image/link as their first message in each group and
thus get their messages marked as suspicious.

Such behavior can be considered spammy, but that's out of the scope of
what this module is defending against.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>