system: Add automatic Git pull-and-restart update command
A convenient update command that performs the equivalent of `git pull`
using GitPython and then initiates a restart. Note that module reloading
is not used here since there may be changes to the core code, which
cannot be reloaded.

The "upd" alias was also stolen from the update_link command since it
fits this new command better.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>