bot: Skip event dispatching path if listener list is empty
If an error occurs during module reloading, there may still be listener
lists present in the listeners dict but they may be empty. Calling
asyncio.wait() on an empty list causes an error, so avoid doing so by
bailing out early when the list is empty.

This fixes the following error:
ValueError: Set of coroutines/Futures is empty.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>