requirements: Remove uvloop
uvloop isn't absolutely necessary for the bot to run. It reduces
portability (doesn't work properly in Termux and doesn't support
Windows) and increases build times for everyone, including those who
don't need the small performance boost that it offers.

Our launcher does fine without it now, so we'll add it as an extra
requirement instead for those who want it.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>