snippet: Bump snippet replacement delay to 1 second
Sometimes some clients are still missing the edit. My theory is that
since Telegram's timestamps are stored in seconds, some clients may be
ignoring the edit update if it arrives at the same time as the intiial
message send update, perhaps due to network delays or cross-DC
interaction.

Whatever the client-side case may be, sleeping for an entire second
should remedy the issue for all since it guarantees that the timestamp
will always be 1 second higher than the original. It does introduce some
visible delay, but that's better than nobody being able to see the
snippet replacement.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>