bot: Allow customizing message response mode
There are now options to either edit the invoking message or reply to it
with the result. Editing is enforced for snippets since that's the way
they're intended to work.

After initially replying to a message, a reference to the reply is saved
so that further calls to msg.result() edit the reply rather than
replying again in order to reduce spam.

Requested-by: XeonDead <xeondead@gmail.com>
Signed-off-by: Danny Lin <danny@kdrag0n.dev>