bot: Add "repost" response mode
This is a new response mode where the bot "reposts" the invocation
message by deleting it and sending a new message that replies to the
same message as the invoker.

While this is fairly spammy and I'm not a fan of it, I can see why some
users might want to use it. However, more importantly, this reduces code
duplication by allowing some commands to enforce this mode for their
file responses.

This also adapts the respond() function to work when text=None to allow
the aforementioned code sharing.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>