Remove PBR in favor of Poetry for packaging and dependency management
Poetry is more modern, flexible, and generally works better than PBR.
For example, unnecessary files (such as Dockerfile) are omitted from
built packages and it supports building wheels. It also enforces version
locking, which is a good practice anyway, so we're using that now.

This also updates the Dockerfile to use Poetry so that everything
continues to work as before.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>