git: Quote all GitPython references in type hints
We support running without GitPython since it's not essential at all
(only used for checking official status for Sentry reporting and the
convenient update command). This is useful for reducing the size of our
Docker image since it's completely unnecessary in that context
(updates are administered by updating the Docker image and users can't
tamper with the code).

Adding type hints breaks startup when the module doesn't load properly
(i.e. when Git isn't installed), so work around the issue by quoting the
type hints so type checkers lazily evaluate them as necessary.

Fixes: 179a34f ("Overhaul with initial type hints and PyCharm inspection fixes")
Signed-off-by: Danny Lin <danny@kdrag0n.dev>