launch, config: Run config upgrader in an asyncio event loop
This will allow us to discard the synchronous database code, which will
be helpful going forward to reduce redundancy and allow an async-only
database in the future.

The upgrader is invoked in its own temporary event loop since it needs
to ensure that the config options necessary for initializing our actual
event loop are present first. We also use aiorun to ensure that all
migration and DB tasks finish, as well as to terminate the program if an
error occurs in the loop during migration and facilitate Ctrl-C
interruption on Windows should the migration process get stuck.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>