telegram_bot: Streamline and extend module event handler registration
New events exposed to modules: MessageDeleted, MessageRead, UserUpdate

Album was omitted because I can't see it being useful for modules.
CallbackQuery and InlineQuery were omitted because they're for bots, not
selfbots. Raw was omitted to avoid exposing Telegram library internals
to modules, though it can be added in the future if necessary.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>