debug: eval: Count errors from the ast module as snippet errors
The built-in `ast` module compiles the snippet to derive an AST (which
meval uses for return insertion), so it encounters syntax errors first.
When thrown, the exception shows the snippet as the "exec" file, but for
some reason that frame isn't visible when we catch it, so check for
ast.py instead. It's a dirty hack, but it's all we can do.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>