context: respond: Implement split/paginate overflow mode
This adds the notion of an "overflow handling mode" to the respond
helper, which can be overridden on a per-command basis, for scenarios
where truncation is not desired. A limit is enforced on the maximum
number of pages per message to help avoid accidental spam.

It's also possible to change the default overflow mode in the config,
but this is discouraged since it generally results in useless spam.

Note that overflow handling is only customizable at the context level as
handling splitting requires some persistent state that's hard to manage
separately.